* Support for mocking long long (>= 64-bit) parameters. * Changed tabs to spaces in cmockery.c/.h to be consistent with the other source files. * Deleted all trailing whitespace. * Fixed warnings introduced in tests after adding support for 64-bit mocks.
git-svn-id: https://cmockery.googlecode.com/svn@37 40f4469a-5155-0410-be90-2de3f0bae501

+++ trunk/src/example/calculator_test.c
@@ -4,9 +4,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -26,8 +26,8 @@
 
 #define array_length(x) (sizeof(x) / sizeof((x)[0]))
 
-/* To simplify this code, these functions and data structures could have been 
- * separated out from the application example.c into a header shared with 
+/* To simplify this code, these functions and data structures could have been
+ * separated out from the application example.c into a header shared with
  * test application.  However, this example illustrates how it's possible to
  * test existing code with little modification. */
 
@@ -48,13 +48,13 @@ extern BinaryOperator find_operator_function_by_string(
         const char* const operator_string);
 extern int perform_operation(
         int number_of_arguments, char *arguments[],
-        const size_t number_of_operator_functions, 
+        const size_t number_of_operator_functions,
         const OperatorFunction * const operator_functions,
         int * const number_of_intermediate_values,
         int ** const intermediate_values, int * const error_occurred);
 extern int example_main(int argc, char *argv[]);
 
-/* A mock fprintf function that checks the value of strings printed to the 
+/* A mock fprintf function that checks the value of strings printed to the
  * standard error stream. */
 int example_test_fprintf(FILE* const file, const char *format, ...) {
 	int return_value;
@@ -69,7 +69,7 @@ int example_test_fprintf(FILE* const file, const char *format, ...) {
 	return return_value;
 }
 
-/* A mock printf function that checks the value of strings printed to the 
+/* A mock printf function that checks the value of strings printed to the
  * standard output stream. */
 int example_test_printf(const char *format, ...) {
 	int return_value;
@@ -120,13 +120,13 @@ void test_divide_by_zero(void **state) {
 	expect_assert_failure(divide(100, 0));
 }
 
-/* Ensure find_operator_function_by_string() asserts when a NULL pointer is 
+/* Ensure find_operator_function_by_string() asserts when a NULL pointer is
  * specified as the table to search. */
 void test_find_operator_function_by_string_null_functions(void **state) {
 	expect_assert_failure(find_operator_function_by_string(1, NULL, "test"));
 }
 
-/* Ensure find_operator_function_by_string() asserts when a NULL pointer is 
+/* Ensure find_operator_function_by_string() asserts when a NULL pointer is
  * specified as the string to search for. */
 void test_find_operator_function_by_string_null_string(void **state) {
 	const OperatorFunction operator_functions[] = {
@@ -136,11 +136,10 @@ void test_find_operator_function_by_string_null_string(void **state) {
 	    array_length(operator_functions), operator_functions, NULL));
 }
 
-/* Ensure find_operator_function_by_string() returns NULL when a NULL pointer 
+/* Ensure find_operator_function_by_string() returns NULL when a NULL pointer
  * is specified as the table to search when the table size is 0. */
 void test_find_operator_function_by_string_valid_null_functions(void **state) {
-  assert_int_equal((int)find_operator_function_by_string(0, NULL, "test"),
-	                     (int)NULL);
+  assert_int_equal(find_operator_function_by_string(0, NULL, "test"), NULL);
 }
 
 /* Ensure find_operator_function_by_string() returns NULL when searching for
@@ -151,9 +150,9 @@ void test_find_operator_function_by_string_not_found(void **state) {
 		{"-", binary_operator},
 		{"/", binary_operator},
 	};
-	assert_int_equal((int)find_operator_function_by_string(
+	assert_int_equal(find_operator_function_by_string(
 	        array_length(operator_functions), operator_functions, "test"),
-	    (int)NULL);
+	        NULL);
 }
 
 /* Ensure find_operator_function_by_string() returns the correct function when
@@ -164,7 +163,7 @@ void test_find_operator_function_by_string_found(void **state) {
 		{"-", (BinaryOperator)0xDEADBEEF},
 		{"/", (BinaryOperator)0xABADCAFE},
 	};
-	assert_int_equal((int)find_operator_function_by_string(
+	assert_int_equal(find_operator_function_by_string(
 	        array_length(operator_functions), operator_functions, "-"),
 	    0xDEADBEEF);
 }
@@ -179,11 +178,11 @@ void test_perform_operation_null_args(void **state) {
 	int error_occurred;
 	expect_assert_failure(perform_operation(
 	    1, NULL, array_length(operator_functions), operator_functions,
-	    &number_of_intermediate_values, &intermediate_values, 
+	    &number_of_intermediate_values, &intermediate_values,
 	    &error_occurred));
 }
 
-/* Ensure perform_operation() asserts when a NULL operator_functions array is 
+/* Ensure perform_operation() asserts when a NULL operator_functions array is
  * specified. */
 void test_perform_operation_null_operator_functions(void **state) {
 	char *args[] = {
@@ -193,11 +192,11 @@ void test_perform_operation_null_operator_functions(void **state) {
 	int *intermediate_values;
 	int error_occurred;
 	expect_assert_failure(perform_operation(
-	    array_length(args), args, 1, NULL, &number_of_intermediate_values, 
+	    array_length(args), args, 1, NULL, &number_of_intermediate_values,
 	    &intermediate_values, &error_occurred));
 }
 
-/* Ensure perform_operation() asserts when a NULL pointer is specified for 
+/* Ensure perform_operation() asserts when a NULL pointer is specified for
  * number_of_intermediate_values. */
 void test_perform_operation_null_number_of_intermediate_values(void **state) {
 	const OperatorFunction operator_functions[] = {
@@ -207,13 +206,13 @@ void test_perform_operation_null_number_of_intermediate_values(void **state) {
 		"1", "+", "2", "*", "4"
 	};
 	int *intermediate_values;
-	int error_occurred;	
+	int error_occurred;
 	expect_assert_failure(perform_operation(
-	    array_length(args), args, 1, operator_functions, NULL, 
+	    array_length(args), args, 1, operator_functions, NULL,
 	    &intermediate_values, &error_occurred));
 }
 
-/* Ensure perform_operation() asserts when a NULL pointer is specified for 
+/* Ensure perform_operation() asserts when a NULL pointer is specified for
  * intermediate_values. */
 void test_perform_operation_null_intermediate_values(void **state) {
 	const OperatorFunction operator_functions[] = {
@@ -223,7 +222,7 @@ void test_perform_operation_null_intermediate_values(void **state) {
 		"1", "+", "2", "*", "4"
 	};
 	int number_of_intermediate_values;
-	int error_occurred;	
+	int error_occurred;
 	expect_assert_failure(perform_operation(
 	    array_length(args), args, array_length(operator_functions),
 	    operator_functions, &number_of_intermediate_values, NULL,
@@ -254,7 +253,7 @@ void test_perform_operation_first_arg_not_integer(void **state) {
 	int *intermediate_values;
 	int error_occurred;
 
-	expect_string(example_test_fprintf, temporary_buffer, 
+	expect_string(example_test_fprintf, temporary_buffer,
 	              "Unable to parse integer from argument test\n");
 
 	assert_int_equal(perform_operation(
@@ -264,7 +263,7 @@ void test_perform_operation_first_arg_not_integer(void **state) {
 	assert_int_equal(error_occurred, 1);
 }
 
-/* Ensure perform_operation() returns an error when parsing an unknown 
+/* Ensure perform_operation() returns an error when parsing an unknown
  * operator. */
 void test_perform_operation_unknown_operator(void **state) {
 	const OperatorFunction operator_functions[] = {
@@ -277,7 +276,7 @@ void test_perform_operation_unknown_operator(void **state) {
 	int *intermediate_values;
 	int error_occurred;
 
-	expect_string(example_test_fprintf, temporary_buffer, 
+	expect_string(example_test_fprintf, temporary_buffer,
 	              "Unknown operator *, argument 1\n");
 
 	assert_int_equal(perform_operation(
@@ -287,20 +286,20 @@ void test_perform_operation_unknown_operator(void **state) {
 	assert_int_equal(error_occurred, 1);
 }
 
-/* Ensure perform_operation() returns an error when nothing follows an 
+/* Ensure perform_operation() returns an error when nothing follows an
  * operator. */
 void test_perform_operation_missing_argument(void **state) {
 	const OperatorFunction operator_functions[] = {
 		{"+", binary_operator},
 	};
 	char *args[] = {
-		"1", "+", 
+		"1", "+",
 	};
 	int number_of_intermediate_values;
 	int *intermediate_values;
 	int error_occurred;
 
-	expect_string(example_test_fprintf, temporary_buffer, 
+	expect_string(example_test_fprintf, temporary_buffer,
 	              "Binary operator + missing argument\n");
 
 	assert_int_equal(perform_operation(
@@ -323,14 +322,14 @@ void test_perform_operation_no_integer_after_operator(void **state) {
 	int *intermediate_values;
 	int error_occurred;
 
-	expect_string(example_test_fprintf, temporary_buffer, 
+	expect_string(example_test_fprintf, temporary_buffer,
 	              "Unable to parse integer test of argument 2\n");
 
 	assert_int_equal(perform_operation(
 	    array_length(args), args, array_length(operator_functions),
 	    operator_functions, &number_of_intermediate_values,
 	    &intermediate_values, &error_occurred), 0);
-	assert_int_equal(error_occurred, 1);	                    
+	assert_int_equal(error_occurred, 1);
 }
